removing some eroneous types throughout and moving towards the Julia Ctype alias typing for better readability when comparing C to julia implementation

+++ src/ExodusConstants.jl
@@ -14,8 +14,8 @@ const IO_word_size = Int32(8) # This was what we did before
 #
 const version_number_1 = Int32(8)
 const version_number_2 = Int32(15)
-const version_number = 8.15
-const version_number_int = 100 * version_number_1 + version_number_2
+const version_number = Float32(8.15)
+const version_number_int = Int32(100 * version_number_1 + version_number_2)
 
 
 # TODO: add the entirity from the exoudsII header
